move application class into its own file
