+++ fink_grb/utils/application.py
@@ -1,0 +1,176 @@
+import sys
+import os
+from enum import Flag, auto
+from dateutil import parser
+
+from astropy.time import Time
+
+import fink_grb
+from fink_grb.offline.spark_offline import spark_offline
+from fink_grb.online.ztf_join_gcn import ztf_join_gcn_stream
+from fink_grb.distribution.distribution import grb_distribution
+
+
+class Application(Flag):
+    OFFLINE = auto()
+    ONLINE = auto()
+    DISTRIBUTION = auto()
+
+    def build_application(self, logger, **kwargs):
+        """ """
+        if self == Application.OFFLINE:
+            application = os.path.join(
+                os.path.dirname(fink_grb.__file__),
+                "offline",
+                "spark_offline.py prod",
+            )
+
+            try:
+                start_window_in_jd = (
+                    Time(parser.parse(kwargs["night"]), format="datetime").jd + 0.49
+                )  # + 0.49 to start the time window in the night
+
+                application += " " + kwargs["hbase_catalog"]
+                application += " " + kwargs["gcn_datapath_prefix"]
+                application += " " + kwargs["grb_datapath_prefix"]
+                application += " " + kwargs["night"]
+                application += " " + str(start_window_in_jd)
+                application += " " + str(kwargs["time_window"])
+                application += " " + kwargs["ast_dist"]
+                application += " " + kwargs["pansstar_dist"]
+                application += " " + kwargs["pansstar_star_score"]
+                application += " " + kwargs["gaia_dist"]
+
+                if kwargs["is_test"]:
+                    application += " " + str(False)
+                else:
+                    application += " " + str(True)
+            except Exception as e:
+                logger.error("Parameter not found \n\t {}\n\t{}".format(e, kwargs))
+                exit(1)
+
+            return application
+        elif self == Application.ONLINE:
+
+            application = os.path.join(
+                os.path.dirname(fink_grb.__file__),
+                "online",
+                "ztf_join_gcn.py prod",
+            )
+
+            try:
+                application += " " + kwargs["ztf_datapath_prefix"]
+                application += " " + kwargs["gcn_datapath_prefix"]
+                application += " " + kwargs["grb_datapath_prefix"]
+                application += " " + kwargs["night"]
+                application += " " + str(kwargs["exit_after"])
+                application += " " + kwargs["tinterval"]
+                application += " " + kwargs["ast_dist"]
+                application += " " + kwargs["pansstar_dist"]
+                application += " " + kwargs["pansstar_star_score"]
+                application += " " + kwargs["gaia_dist"]
+            except Exception as e:
+                logger.error("Parameter not found \n\t {}\n\t{}".format(e, kwargs))
+                exit(1)
+
+            return application
+
+        elif self == Application.DISTRIBUTION:
+
+            application = os.path.join(
+                os.path.dirname(fink_grb.__file__),
+                "distribution",
+                "distribution.py prod",
+            )
+
+            try:
+                application += " " + kwargs["grb_datapath_prefix"]
+                application += " " + kwargs["night"]
+                application += " " + str(kwargs["exit_after"])
+                application += " " + kwargs["tinterval"]
+                application += " " + kwargs["kafka_broker"]
+                application += " " + kwargs["username_writer"]
+                application += " " + kwargs["password_writer"]
+            except Exception as e:
+                logger.error("Parameter not found \n\t {}\n\t{}".format(e, kwargs))
+                exit(1)
+
+            return application
+
+    def run_application(self):
+        if self == Application.OFFLINE:
+
+            hbase_catalog = sys.argv[2]
+            gcn_datapath_prefix = sys.argv[3]
+            grb_datapath_prefix = sys.argv[4]
+            night = sys.argv[5]
+            start_window = float(sys.argv[6])
+            time_window = int(sys.argv[7])
+
+            ast_dist = float(sys.argv[8])
+            pansstar_dist = float(sys.argv[9])
+            pansstar_star_score = float(sys.argv[10])
+            gaia_dist = float(sys.argv[11])
+
+            column_filter = True if sys.argv[12] == "True" else False
+
+            spark_offline(
+                hbase_catalog,
+                gcn_datapath_prefix,
+                grb_datapath_prefix,
+                night,
+                start_window,
+                time_window,
+                ast_dist,
+                pansstar_dist,
+                pansstar_star_score,
+                gaia_dist,
+                with_columns_filter=column_filter,
+            )
+
+        elif self == Application.ONLINE:
+
+            ztf_datapath_prefix = sys.argv[2]
+            gcn_datapath_prefix = sys.argv[3]
+            grb_datapath_prefix = sys.argv[4]
+            night = sys.argv[5]
+            exit_after = sys.argv[6]
+            tinterval = sys.argv[7]
+
+            ast_dist = float(sys.argv[8])
+            pansstar_dist = float(sys.argv[9])
+            pansstar_star_score = float(sys.argv[10])
+            gaia_dist = float(sys.argv[11])
+
+            ztf_join_gcn_stream(
+                ztf_datapath_prefix,
+                gcn_datapath_prefix,
+                grb_datapath_prefix,
+                night,
+                exit_after,
+                tinterval,
+                ast_dist,
+                pansstar_dist,
+                pansstar_star_score,
+                gaia_dist,
+            )
+
+        elif self == Application.DISTRIBUTION:
+
+            grbdata_path = sys.argv[2]
+            night = sys.argv[3]
+            exit_after = sys.argv[4]
+            tinterval = sys.argv[5]
+            kafka_broker = sys.argv[6]
+            username_writer = sys.argv[7]
+            password_writer = sys.argv[8]
+
+            grb_distribution(
+                grbdata_path,
+                night,
+                tinterval,
+                exit_after,
+                kafka_broker,
+                username_writer,
+                password_writer,
+            )
